Crate ogg_metadata

Metadata parser library for various Ogg formats.

Supported formats:

Vorbis (Detect, Metadata)

Opus (Detect, Metadata)

Theora (Detect, Metadata)

Speex (Detect)

Support will be extended in the future, especially for the Theora codec.

Modules

opus

theora

vorbis

Structs

OpusMetadata
Metadata for the Opus audio codec.

TheoraMetadata
Metadata for the Theora video codec.

VorbisMetadata
Metadata for the Vorbis audio codec.

Enums

BareOggFormat
Bare (C-style enum) counterpart to OggFormat

OggFormat

OggMetadataError

Traits

AudioMetadata

Functions

format_duration

get_absgp_of_last_packet

identify_packet_data_by_magic

needs_last_packet_absgp
Returns whether the rich metadata has lenth information, which requires retrieving of the absgp of the last packet.

parse_format

read_format
Reads the format of the file.

seek_before_end